refactor: modernize Swift style
